% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{efftox_analysis_to_df}
\alias{efftox_analysis_to_df}
\title{EffTox analysis to data.frame}
\usage{
efftox_analysis_to_df(x)
}
\arguments{
\item{x}{An \code{\link{efftox_analysis}}}
}
\value{
a \code{data.frame}
}
\description{
Convenient function to turn an \code{\link{efftox_analysis}}
into a \code{data.frame}.
}
\examples{
design <- efftox_parameters_demo()
design$n <- 3
design$yE <- c(0, 1, 1)
design$yT <- c(0, 0, 1)
design$levels <- c(1, 2, 3)
fit <- rstan::sampling(stanmodels$EffTox, data = as.list(design))
decision <- efftox_process(design, fit)
df <- efftox_analysis_to_df(decision)
round(df$Utility, 2) == c(-0.64, 0.04, 0.24, -0.05, -0.19)

}
