% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{efftox_process}
\alias{efftox_process}
\title{Process RStan samples from an EffTox model}
\usage{
efftox_process(design, fit)
}
\arguments{
\item{design}{An instance of \code{\link{efftox_params}}, a list of EffTox
parameters. An example is yielded by \code{\link{efftox_parameters_demo}}.}

\item{fit}{An instance of \code{rstan::stanmodel}, derived by sampling an
EffTox model.}
}
\value{
An instance of \code{\link{efftox_analysis}}.
}
\description{
Process RStan samples from an EffTox model to make inferences
about dose-acceptability, dose-utility and which dose should be recommended
next.
}
\examples{
design <- efftox_parameters_demo()
design$n <- 3
design$yE <- c(0, 1, 1)
design$yT <- c(0, 0, 1)
design$levels <- c(1, 2, 3)
fit <- rstan::sampling(stanmodels$EffTox, data = as.list(design))
decision <- efftox_process(design, fit)
decision$recommended_dose == 3
}
\seealso{
\code{\link{efftox_params}}

\code{\link{efftox_parameters_demo}}
}
