% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{efftox_theta}
\alias{efftox_theta}
\title{Estimate EffTox logit model coefficients}
\usage{
efftox_theta(probs, doses, degree = 1, log = TRUE)
}
\arguments{
\item{probs}{vector of probabilites at which to estimate coefficients.}

\item{doses}{vector of corresponding doses.}

\item{degree}{degree of the polynomial dose effect in the logit model.}

\item{log}{logical indicating whether doses are to be log transformed and
centered in the analysis.}
}
\value{
A vector of coefficients.
}
\description{
Estimate regression coefficients for the efficacy or toxicity logit models in
an EffTox trial design, given a set of probabilities and doses.
}
\seealso{
\code{\link{EffToxDesign}}
}
