% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{efftox_utility_density_plot}
\alias{efftox_utility_density_plot}
\title{Plot densities of EffTox dose utilities}
\usage{
efftox_utility_density_plot(fit, doses = NULL)
}
\arguments{
\item{fit}{An instance of \code{rstan::stanmodel}, derived by sampling an
EffTox model.}

\item{doses}{optional, vector of integer dose-levels to plot. E.g. to plot
only dose-levels 1, 2 & 3 (and suppress the plotting of any other doses), use
\code{doses = 1:3}}
}
\value{
an instance of \code{ggplot}. Omit assignment to just view the plot.
}
\description{
Plot densities of EffTox dose utilities. Optionally plot only a
subset of the doses by specifying the \code{doses} parameter. This function
requires ggplot2 be installed.
}
\note{
This function requires that ggplot2 be installed.
}
\examples{
design <- efftox_parameters_demo()
design$n <- 3
design$yE <- c(0, 1, 1)
design$yT <- c(0, 0, 1)
design$levels <- c(1, 2, 3)
fit <- rstan::sampling(stanmodels$EffTox, data = as.list(design))
efftox_utility_density_plot(fit) + ggplot2::ggtitle('My doses')  # Too busy?
# Specify subset of doses to make plot less cluttered
efftox_utility_density_plot(fit, doses = 1:3) + ggplot2::ggtitle('My doses')

}
